Bug fixes, additional debug code, extended functionality
SoapForwardHub should now wrap all elements it should send in Envelopes.

Converting a W3CEndpointReference to a String via getAddress in
ServiceUtilities now produces a lot more debug output.

Concrete and Simple evaluators should now accept topic expressions
starting with "/", even though this is not allowed in OASIS specification.

A sendSingleNotify() has been added to AbstractNotificationProducer, which
does not store the Notify for later retrieval, and does not require the
target address to be registered beforehand. It will not register the
address.

GenericNotificationProducer got constructors which sets its Hub.