Mainly filling in exceptions where they are thrown
A LOT were just empty, causing erronous data to be returned from Web
Service. Some of them are now filled in, revealing other faults -> new
branch.

Fixed a bug in the defaultconfig.xml where references to objects was not
properly referenced.

Moved bindings of hub in quickbuild in some of our implementations, so the
warning about endpoint set before hub was initialized dissapeared.